Fixed handling of ctrl-t
Before, when exercising the quit key sequence (ctrl-t + q) the ctrl-t code
(0x14) would be sent.

This is now fixed so that it is not sent.

However, in case it is needed to send ctrl-t to the device it is possible by
simply repeating the ctrl-t.

Meaning, ctrl-t + ctrl-t = ctrl-t sent to device.